XD-886 Fix package tangle - Moved DistributedJobService to dirt/plugins - Moved ExpandedJobInfo to dirt/job
